+++ cert/gen.sh
@@ -11,4 +11,4 @@ fi
 
 rm -f *.pem
 
-openssl req -x509 -newkey rsa:4096 -days $DAYS -nodes -keyout server-key.pem -out server-cert.pem -subj "/O=$SERVER_NAME/CN=$SERVER_IP" -addext "subjectAltName=IP:$SERVER_IP"
+openssl req -x509 -newkey rsa:4096 -days $DAYS -nodes -keyout server-key.pem -out server-cert.pem -subj "/O=$SERVER_NAME/CN=$SERVER_PUBLIC_IP" -addext "subjectAltName=IP:$SERVER_PUBLIC_IP"